+++ mods/src/misc/api.cpp
@@ -16,7 +16,7 @@ struct Callbacks {
     std::vector<std::function<void(Args...)>> functions;
     void run(Args... args) {
         for (const std::function<void(Args...)> &func : functions) {
-            func(args...);
+            func(std::forward<Args>(args)...);
         }
     }
 };